with_poetry/pyproject.toml
[tool.poetry]
name = "with-poetry"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "with_poetry"}]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^7.2.0"
selene = "^1.0.2"
allure-pytest = "^2.12.0"
requests = "^2.28.1"
testrail = "^0.3.15"
voluptuous = "^0.13.1"
pytest-voluptuous = "^1.2.0"
curlify = "^2.2.1"



[tool.pytest.ini_options]

addopts = "--alluredir allure-results --clean-alluredir"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
